+++ src/index.js
@@ -1,4 +1,6 @@
-module.exports = {
-  Plugin: require('./plugin'),
-  SearchProvider: require('./search-lunr'),
-};
+import SearchProvider from './search-lunr.js';
+
+export default SearchProvider;
+
+export { default as SearchProvider } from './search-lunr.js';
+export { default as Plugin } from './plugin.js';