+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [4.0.0](https://github.com/uttori/uttori-search-provider-lunr/compare/v3.3.5...v4.0.0) - 2023-12-26
+
+- 💥 BREAKING CHANGES!
+- 💥 Rename config key: `ignore_slugs` to `ignoreSlugs`
+- 💥 Add required config for `lunr_locales` of `lunrLocaleFunctions` to be an array of lunr locale functions rather than strings
+- 🎁 Convert to be a module, and require Node v20
+- 🎁 Update dependencies
+- 🎁 Update dev dependencies
+- 🛠 Update ESLint configuration
+- 🛠 Update NVM to v20.10.0
+- 🛠 Update old tooling configuration files
+- 🛠 Remove CJS support, restructure to export ESM only
+
 ## [3.3.5](https://github.com/uttori/uttori-search-provider-lunr/compare/v3.3.4...v3.3.5) - 2021-02-28
 
 - 🎁 Update dependencies